Use mbostock/shapefile for reading shapefiles
I was having problems installing the shp2json dependency
(https://www.npmjs.com/package/shp2json) because of its dependency
on the gdal package. There weren't prebuilt binaries for the current
stable version of node and the required version of gdal and the
compilation kept failing.

I decided to use mbostock/shapefile which is a bit newer and doesn't
depend on gdal.

However, using this library means I have to handle unzipping the
shapefile myself, which adds some dependencies.

I also replace the use of the request package as it has been deprecated.
I chose to use node-fetch instead.

Addresses #2
https://github.com/ghing/tiger-boundaries/issues/2